Atualizar o route.ts

+++ app/api/admin/desafios/[id]/route.ts
@@ -31,8 +31,8 @@ export async function PATCH(
         if (valor !== undefined) dataToUpdate.valor = parseFloat(valor);
         if (duracaoDias !== undefined) dataToUpdate.duracaoDias = parseInt(duracaoDias) || 90;
         if (dataInicio !== undefined) dataToUpdate.dataInicio = dataInicio ? new Date(dataInicio) : null;
-        if (pontosPorItem !== undefined) dataToUpdate.pontosPorItem = parseInt(pontosPorItem) || 1;
-        if (pontosPorItemFimDeSemana !== undefined) dataToUpdate.pontosPorItemFimDeSemana = parseInt(pontosPorItemFimDeSemana) || 2;
+        if (pontosPorItem !== undefined) dataToUpdate.pontosPorItem = parseFloat(pontosPorItem) || 1;
+        if (pontosPorItemFimDeSemana !== undefined) dataToUpdate.pontosPorItemFimDeSemana = parseFloat(pontosPorItemFimDeSemana) || 2;
         if (linkGrupoWhats !== undefined) dataToUpdate.linkGrupoWhats = linkGrupoWhats;
         if (ativo !== undefined) dataToUpdate.ativo = ativo;
         // 🔑 coachId agora é editável também depois de criado — define de
